keep only one copy of file handling functions
