+++ core-client/Cargo.toml
@@ -60,6 +60,7 @@ serde_json.workspace = true
 strum.workspace = true
 strum_macros.workspace = true
 tempfile.workspace = true
+test-utils.workspace = true
 tfhe.workspace = true
 threshold-networking.workspace = true
 tokio = { workspace = true, features = ["rt-multi-thread"] }
@@ -75,7 +76,6 @@ validator.workspace = true
 # Use .workspace = true to reference workspace dependencies
 # Maintain alphabetical order
 assert_cmd.workspace = true
-test-utils.workspace = true
 test-utils-cc = { path = "./test-utils-cc" }
 futures.workspace = true
 kms = { workspace = true, features = ["insecure", "testing"] }
