+++ core/service/src/client/mod.rs
@@ -12,8 +12,6 @@ pub mod key_gen;
 #[cfg(feature = "non-wasm")]
 pub mod local_crypto;
 #[cfg(feature = "non-wasm")]
-pub mod local_files;
-#[cfg(feature = "non-wasm")]
 pub mod mpc_context;
 #[cfg(feature = "non-wasm")]
 pub mod public_decryption;
