+++ core/test-utils/Cargo.toml
@@ -5,11 +5,12 @@ edition.workspace = true
 publish.workspace = true
 authors.workspace = true
 license.workspace = true
-description = "Testing utilities"
+description = "Shared testing and local file utilities"
 
 [dependencies]
 anyhow.workspace = true
 bc2wrap.workspace = true
 serde.workspace = true
+tempfile.workspace = true
 tokio.workspace = true
 tracing-subscriber = { workspace = true, features = ["env-filter", "fmt"] }
